requirements.txt
requests>=2.32.0
langchain-openai>=0.3.0
langgraph>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.1
tavily-python>=0.5.0
yfinance>=0.2.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
